+++ src/app/services/ceph-manager.js
@@ -167,8 +167,8 @@ export default Service.extend({
   },
 
   /**
-   * Checks whether local ceph storage can be created or not
-   * @returns {Promise<boolean>} resolves to true if ceph local storage
+   * Checks whether embedded ceph storage can be created or not
+   * @returns {Promise<boolean>} resolves to true if ceph embedded storage
    *   can be created
    */
   canCreateStorage() {