+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-8800 Renamed "Local Ceph" storage type to "Embedded Ceph"
 * VFS-7272 Changed let's encrypt form
 * VFS-8568 Added automation main menu item
 * VFS-8617 Removed usages of ember-invoke-action